create prettyFromJsonStr out of the previous mess
still a mess, but more workable
