+++ jevkoToPrettyString.js
@@ -1,12 +1,11 @@
-import {parseJevkoWithHeredocs} from "https://cdn.jsdelivr.net/gh/jevko/parsejevko.js@v0.1.8/mod.js"
 import {escape} from 'https://cdn.jsdelivr.net/gh/jevko/jevkoutils.js@v0.1.6/mod.js'
 
-const sth = (str, tag) => `\`/${tag}/${str}/${tag}/`
+const strToHeredoc = (str, tag) => `\`/${tag}/${str}/${tag}/`
 
-const jevkoToPrettyString = (jevko) => {
+export const jevkoToPrettyString = (jevko) => {
   const {subjevkos, suffix, tag} = jevko
 
-  if (tag !== undefined) return sth(suffix, tag)
+  if (tag !== undefined) return strToHeredoc(suffix, tag)
 
   let ret = ''
   for (const {prefix, jevko} of subjevkos) {
@@ -17,10 +16,11 @@ const jevkoToPrettyString = (jevko) => {
 
 const escapePrefix = (prefix) => prefix === ''? '': escape(prefix) + ' '
 
+//?todo: unhardcode []`
 const recur = (jevko, indent, prevIndent) => {
   const {subjevkos, suffix, tag} = jevko
 
-  if (tag !== undefined) return sth(suffix, tag)
+  if (tag !== undefined) return strToHeredoc(suffix, tag)
 
   let ret = ''
   if (subjevkos.length > 0) {
@@ -35,7 +35,7 @@ const recur = (jevko, indent, prevIndent) => {
   return '[' + ret + escape(suffix) + ']'
 }
 
-const parsed = parseJevkoWithHeredocs(Deno.readTextFileSync('test.jevkodata'))
+// const parsed = parseJevkoWithHeredocs(Deno.readTextFileSync('test.jevkodata'))
 
-// todo: heredoc-aware pretty printer
-console.log(jevkoToPrettyString(parsed))
+// // todo: heredoc-aware pretty printer
+// console.log(jevkoToPrettyString(parsed))
